+++ packages/terra-data-grid/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added fromWorkListDataGrid a private prop that gets passed to terra-table.
+  * Added rowMinimumHeight a private prop that gets passed to terra-table.
 
 ## 1.11.0 - (January 12, 2024)